+++ BloodStore/Assets/BloodPack.json
@@ -1,0 +1,1 @@
+{"arr":[{"node":{"name":"David","sex":"Male","bloodType":["A","-","AO"],"hp":100,"age":35,"isDead":false,"empty":true},"num":4,"date":1,"existNext":false},{"node":{"name":"Frank","sex":"Female","bloodType":["AB","-","AB"],"hp":100,"age":32,"isDead":false,"empty":true},"num":2,"date":1,"existNext":false}]}
