+++ BloodStore/Assets/FamilyTree.json
@@ -1,0 +1,1 @@
+{"arr":[{"male":{"name":"David","sex":"Male","bloodType":["A","-","AO"],"hp":100,"age":35,"isDead":false,"empty":false},"female":{"name":"Frank","sex":"Female","bloodType":["AB","-","AB"],"hp":100,"age":32,"isDead":false,"empty":false},"isPair":true,"childNum":3},{"male":{"name":"Charlie","sex":"Male","bloodType":["AB","-","AB"],"hp":50,"age":8,"isDead":false,"empty":false},"female":{"name":"","sex":"","bloodType":[],"hp":0,"age":0,"isDead":false,"empty":true},"isPair":false,"childNum":0},{"male":{"name":"","sex":"","bloodType":[],"hp":0,"age":0,"isDead":false,"empty":true},"female":{"name":"Frank","sex":"Female","bloodType":["A","-","AA"],"hp":50,"age":4,"isDead":false,"empty":false},"isPair":false,"childNum":0},{"male":{"name":"","sex":"","bloodType":[],"hp":0,"age":0,"isDead":false,"empty":true},"female":{"name":"Hank","sex":"Female","bloodType":["A","-","AA"],"hp":50,"age":6,"isDead":false,"empty":false},"isPair":false,"childNum":0}]}
