+++ BloodStore/Assets/FamilyTree.json.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 49ce4c4a768573249a7e521ad26add58
-PrefabImporter:
+guid: 428e6bfd57e9c18488e98bb7015d1a2d
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
