--- BloodStore/Assets/Prefab/Card.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d047bfadbc49fe54a930ef1b474f6fd9
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
