--- BloodStore/Assets/Prefab/Group.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 52b2e7a1eeb50974b84c88b57f33e6f9
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
