--- BloodStore/Assets/Prefab/Node.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d2072b4a70e599849bb7d21195f029bb
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
