+++ BloodStore/Assets/Scripts/LeeHyangWoo/NewFamilyTree/RightFuture/BloodPacks.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e96fbca96571764a96d2736a9382aa3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
